Renamed the page objects
