+++ page_objects/examples/weather_shopper_mobile_app/weather_shopper_home_page.py
@@ -5,7 +5,7 @@ This class models the home page in Weathershopper application.
 import conf.locators_conf as locators
 from utils.Wrapit import Wrapit
 from core_helpers.mobile_app_helper import Mobile_App_Helper
-from .weather_shopper_homepage_objects import HomepageObjects
+from .homepage_objects import HomepageObjects
 from .navigation_menu_objects import NavigationMenuObjects
 
 class WeatherShopperHomePage(Mobile_App_Helper, HomepageObjects, NavigationMenuObjects):
